feat(server): store the context restore outcome from the perf beacon
The client beacon now sends contextRestoredCount and
contextRestoreFailures beside contextLostCount (issue 3846). Parse both
with the same clamp, persist them in two additive client_perf_reports
columns, surface them on the admin raw rows and sum the failures into
the perf summary as contextRestoreFailureCount. The Prometheus context
loss series is untouched.

Adding the columns pushed server/db.ts past its ratchet ceiling, so the
whole client perf domain (table DDL, insert, retention prune) moves to
server/client_perf_db.ts with its own CLIENT_PERF_SCHEMA applied right
after the core SCHEMA, reaching the pool through a memoized lazy import
because db.ts imports the schema at module scope. The db.ts ceiling is
lowered to the new count.
